// src/models/enum/grupoAlimenticio.ts
// Definición del enum GrupoAlimenticio
export enum GrupoAlimenticio {
  Lacteos = "Lácteos y derivados",
  Huevos = "Huevos y derivados",
  Carnicos = "Cárnicos y derivados",
  Pescados = "Pescados, moluscos, reptiles, crustáceos y derivados",
  Grasas = "Grasas y aceites",
  Cereales = "Cereales y derivados",
  Legumbres = "Legumbres, semillas, frutos secos y derivados",
  Verduras = "Verduras, hortalizas y derivados",
  Frutas = "Frutas y derivados",
  Azucar = "Azúcar, chocolate y derivados",
  Bebidas = "Bebidas (no lácteas)",
  Miscelanea = "Miscelánea",
  Otro = "Otro",
}