// src/models/enum/alergeno.ts
// Definición del enum Alergeno
export enum Alergeno {
  Cereales = "Cereales con gluten",
  Crustaceos = "Crustáceos y productos a base de crustáceos",
  Huevos = "Huevos y productos derivados",
  Pescado = "Pescado y productos a base de pescados",
  CacahuetesFrutosSecos = "Cacahuetes, productos a base de cacahuetes y frutos secos",
  Soja = "Soja y productos a base de soja",
  Leche = "Leche y sus derivados (incluida la lactosa)",
  FrutosCascara = "Frutos de cáscara y productos derivados",
  Apio = "Apio y productos derivados",
  Mostaza = "Mostaza y productos a base de mostaza",
  Sesamo = "Granos o semillas de sésamo y productos a base de sésamo",
  DioxidoAzufre = "Dióxido de azufre y sulfitos",
  Altramuces = "Altramuces y productos a base de altramuces",
  Moluscos = "Moluscos y crustáceos y productos a base de estos",
}