// src/models/enum/incompatibilidadAlimenticia.ts
// Definición del enum IncompatibilidadAlimenticia
export enum IncompatibilidadAlimenticia {
  Celiacos = "No adecuado para celiacos", // (gluten)
  Diabeticos = "No adecuado para diabeticos", // (alto contenido de azúcares o carbohidratos)
  Semivegetarianos = "No adecuado para semivegetarianos", // (contiene carne o pescado)
  Vegetarianos = "No adecuado para vegetarianos", // (contiene carne o pescado)
  Veganos = "No adecuado para veganos", // (contiene ingredientes de origen animal)
}